[IMP] ai: implementation of a search topic
This commit introduce a new topic "Information retrieval" which can be
used to ask the LLM to give/analyze informations present in the
database.

To do that it uses the available models (models visible by the user,
that are installed), and an already existing tool `get_fields` which
allows to retrieve the fields of a model.

This commit also introduce two new tools that are essential for the
information retrieval:
- `search`: Which can be used by the agent to perform a `search_read`
call on a model,
- `read_group`: which exposes the `_read_group` method as well.

task-5028508

closes odoo/enterprise#93041

Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>