[FIX] mrp_workorder: display app title on reload
**PROBLEM**
In debug mode, we can change the default home action of a user (the action he sees when logging in). When the action `action_mrp_display` is set as the home action, there is a traceback after logging in.

**STEP TO REPRODUCE**
1. Go in debug mode
2. Change the home action of a user to the 'Shop Floor' action (in the user form, in the preference tab).
3. log out, and log in with this user.
4. a js traceback should appear.

**CAUSE**
In mrp_workcenter_dialog.js, the Shop Floor action uses the `menu` service to get the name of the current app.
`setCurrentMenu()` which set the current app in the ̀`menu` service is not called before the `appName` getter is called.
This lead to a traceback because this.menu.getCurrentApp() is `undefined`.

**FIX**
Workaround if the current app is undefined.
We already check what is returned by `getCurrentApp()` where it is used.

opw-4926317

closes odoo/enterprise#94957

X-original-commit: 07a4d4832f6fbdbb57b3892757f28bac5c1703e4
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>
Signed-off-by: Léo Gizard (legi) <legi@odoo.com>