[FIX] l10n_cl_edi_pos: avoid crash upon order validation
Currently, when having a CL Company an dusing the POS, it
is possible to have the POS crash when validating orders.

Steps to reproduce:
-------------------
* Install l10n_cl_edi_pos
* Create a new company
* Put Chile as country
* Switch to that company
* Create a shop and open it
* Make a sale, validate order
> Observation: POS Crashes

Why the fix:
------------
The crash is happening upon trying to render the receipt because
`l10n_cl_dte_resolution_date` is undefined.

`l10n_cl_dte_resolution_date`is a field that can be found on the
company settings only if the field `l10n_cl_dte_service_provider` is
set and is required when the later has either `SII` or `SIITEST` as
value.

For companies in Chile, there is therefore two setups that leads to
having `l10n_cl_dte_resolution_date` undefined. Either the service
provider is set to demo, which case clients don't need to fill in the
resolution date, either the provider is left empty (since not required
upon company creation).

Since the field can be undefined, we don't render it when it is.

opw-5034091

closes odoo/enterprise#93930

X-original-commit: d572cfe3cda5668d413e5bf582cab35de4203522
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>