[IMP] voip_ai: Generate and display call's transcript summary
This commit introduces the functionality to generate a one-liner summary
for VoIP call transcripts.

We also skip adding datetime to the generated transcript as it is no
longer required.

Part of the followup to the commit introducing transcriptions:
f96d1f4cc52c009a4f55cfaabf7b683bdf9e170d
[ADD] ai_voip: end-to-end post-call recording & transcription

Related to
task-4532108

Part-of: odoo/enterprise#92954
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>