[FIX] quality_iot: multiple cameras not working
Due to an issue with the `CameraInterface`, the same identifier would be
chosen for any connected camera (`camera-2`), meaning only one camera
could be used at once.

To fix this, the `camera.id` property is used as the identifier instead.
The `libcamera` documentation says the following:

> The camera ID is a free-form string that identifies a camera in the
> system. IDs are guaranteed to be unique and stable: the same camera,
> when connected to the system in the same way
> (e.g. in the same USB port), will have the same ID across both
> unplug/replug and system reboots.

However, we will also still return the last camera with the `camera-2`
identifier, so that existing device configurations in the backend don't
get lost. In master we will remove this legacy identifier.

In addition, another small fix was made to the `CameraDriver` to ensure
a string is sent rather than raw bytes, this prevents a traceback when
communicating over the websocket.

task-4792552

closes odoo/enterprise#93148

X-original-commit: cc490083aa6f459495dc896aea2c79a17d7bef0c
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>