[IMP] voip_ai: Call from view shows transcription_status as a banner
Before this commit transcription_status was presented as a selection field
on the voip.call form view.

With this commit we move this information out of there and put it in a
banner that only shows up when transcription is in progress or in error.

related tasks
task-5026049

Part-of: odoo/enterprise#92954
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>