[FIX] ai_documents: fix crash after previous tool move
Bug
===
If the LLM executes an action (e.g. "create vendor bill") that moves the
document by side effect and then executes the action "move document",
an error is displayed in the chatter.

The new folder won't have a prompt, and the check in
`_ai_action_move_in_folder` will fail.

Instead of relying on the document's current folder,
we should try to get the folder linked to the action
and not to the document.

Task-4989962

closes odoo/enterprise#91601

Related: odoo/odoo#223159
Related: odoo/upgrade#8157
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>