[FIX] l10n_my_reports: currency not converted in aged receivable report
When we create an invoice with a different currency than the main one,
the Statement of Account PDF report in aged receivable has an error.
The amount for each line is displayed as the main currency, but the
conversion isn't done.
For instance, if we create an invoice line of 3.5$ but our main
currency is the Euro, in the report, it will show 3.5€ instead of 2.99€

1. In a db without demo data, Download the app l10n_my.
2. In the general settings, click on Currencies. Activate a second
currency and set its currency rate (relative to the main currency of
the company) different of 1.
3. In accounting > settings, click on -> Currencies. Activate this
second currency.
4. Click on configuration > journals and activate the debug mode. In
the Journal Entries tab, as Currency, select a different currency than
your main one (set in the general settings).
5. Create an invoice. Set a customer and select the journal for
which you set an additional currency. Select this additional currency.
Finally, add a product and click confirm.
6. Click on Reporting > Aged Receivable. Then, click on the line
corresponding to your partner on Statement of Account. A PDF should
download.

The lines under Balance should be displayed as the main currency of the
company, with the correct currency rates applying.

The lines under Balance are displayed as the main currency of the
company but the currency rates are not applying.

In the code, the currency rate isn't used to generate this line of
the report.
__
opw-4975540

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr

closes odoo/enterprise#93767

X-original-commit: 8c53e50df1cf9dc6d3ca4cae19c39135ac85d4e4
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>