[IMP] l10n_dk_report: EC sales list and tax report wizard
Following the implementation of the Tax returns feature in 18.3 which
was focused on the general implementation and Belgium, we want to add
the specification for other country and in this case Denmark.

EC Sales List is a part of this, with VAT Returns as the EC Sales List
is generated when the VAT Return is submitted.

We are showing modals with the instructions as well as a Download CSV
button to make it easier for them to have access to the CSV since it's
important for the users to have the instructions present when
submitting those documents.

This also includes the deadlines and frequencies for the generation of
those documents, and also the translation of the instructions.

For the tax report, since Denmark has an api to send the report. When submitting
the closing, the user will be redirected to the wizard to send it.

task-4826286

closes odoo/enterprise#93223

Related: odoo/odoo#225011
Related: odoo/upgrade#8298
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>