[IMP] hr_work_entry_enterprise: sum of hours per rows and col
Problem
----------
Do the same row sum computation as in planning.

Objective
----------
per employee X/h / Y/h (%)
X: sum of work entry durations
Y: sum of theoretical hours per day

per column : sum of durations

Solution
----------
Do the same behaviour as work entry generation.
This raises issues in the work entry generation. attendance_ids are retrieved from the working schedule with `_get_attendance_intervals` method per intervals and not per versions. It makes the computation longer and incorrect in the case of a new version with a flexible working schedule on an employee. It will not be considered as flexible.

task-5039278

closes odoo/enterprise#93835

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>