[FIX] l10n_mx_edi_pos: prevent refund excess
When refunding an order that originally had a global discount, if you
didn't refund the discount, the refund total amount would exceed the
original order total. This would cause issues when trying to generate
a global invoice for the mexican localization.

Steps to reproduce:
-------------------
* Activate the global discount option in any PoS
* Open PoS and make a sale with a global discount
* Refund the sale without including the discount
* Go to the backend
* Go to the order list and select the 2 orders you made
* Now try to create a global invoice
> Observation: The global invoice is in error because the negative lines
  cannot be distributed correctly.

Why the fix:
------------
To avoid this issue with the global invoice we simply prevent the user
to generate a refund with a greater amount than the original order. We
only apply this limit to the mexican localization because it's the only
module that is affected by this issue. Other localizations can still
refund without restrictions even if this work flow does not really make
sense.

opw-4899501

closes odoo/enterprise#101752

X-original-commit: 2813643e4bf8270264e38e6fc5f390e4bebcf4ed
Signed-off-by: David Monnom (moda) <moda@odoo.com>
Signed-off-by: Robin Engels (roen) <roen@odoo.com>