[FIX] delivery_dhl_rest: fixing insurance percentage not taken into account
**PROBLEM**
Insuring a delivery using the dhl carrier don't work.
It doesn't affect the estimated rate of the delivery, and the shipment created when validating the delivery order isn't insured.

**STEP TO REPRODUCE**
1. Install the `delivery_dhl_rest` and the `l10n_be` modules (we will use the be demo company).
2. Set the insurance percentage of the dhl be delivery method to 100%, and set the region to Europe (the demo data is incorrect), and activate the debug (click the "No Debug" smart button to activate the log of requests).
3. Switch to the be company.
5. Create a sale order, with a customer located in Belgium, and add shipping using the dhl method.
6. Go to the delivery order, and validate it.
7. Go to settings/Technical/Logging and look at the rating_request and shipment_request, notice there is no information about insurance.

**CAUSE**
We don't send any info about insurance in the api requests.

**FIX**
Computing and sending the insured amount, only if the insurance percentage is not null.
If the package can't be insured between the origin and the destination, a error message will be displayed when updating the delivery price.

opw-4989281

X-original-commit: 403d4f4b009a121bbfc3dab5916cba1c17d7285e
Part-of: odoo/enterprise#96210
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Léo Gizard (legi) <legi@odoo.com>