[IMP] l10n_br_avatax, l10n_br_edi: add CFOP Field on operation type
We allow the user to set a custom CFOP on l10n_br.operation.type.

This allows the to set a global CFOP on a record (account.move or
sale.order) and also to override that per line through the existing
l10n_br_goods_operation_type_id m2o fields on the parent and line
models.

Because it doesn't affect tax calculation we limit the change to just
EDI.

Our general approach is to do tax calculation and save the response in
`l10n_br_edi_avatax_data` to resubmit it for EDI (the EDI request is
essentially a superset of the tax calculation response). This approach
doesn't work here because they keep returning the CFOP they determine,
not the one we send. Therefore we patch the place where we save the
tax calculation response to add the right CFOPs.

Task-4947957

closes odoo/enterprise#93240

Signed-off-by: Laurent Smet (las) <las@odoo.com>
Co-authored-by: hisi-odoo <hisi@odoo.com>