[IMP] web_enterprise,*: light/dark mode persistence & browser pref.
*: test_discuss_full_enterprise

This commit implements the support for the browser/device/system
prefered color scheme in addition to arbitrary light/dark schemes.

Also, it persists accross sessions (and devices) the color scheme
choosen by the user in a `res.user.settings`.

Note that, as for the existing "dark mode", it only applies to the
backend.

About the browser/device/system scheme support, please note that as the
only current way to actually detect the color scheme (light/dark) to
apply is via the `prefers-color-scheme` media query, hence run in the
browser ; the server as no way to know it by itself [1]. To work around
this limitation, the (existing) cookie is used as the link between the
client and server. The drawback is that a reload has to be performed
when the user has its browser/device/system set to "dark" and connects
for the first time.

[1]: Chrome implements the HTTP header `Sec-CH-Prefers-Color-Scheme` but
no other browser does at this time and the corresponding spec is very
much a "work in progress/draft" at the moment...

References:
- https://developer.mozilla.org/en-US/docs/Web/CSS/@media/prefers-color-scheme
- https://chromestatus.com/feature/5642300464037888
- https://wicg.github.io/user-preference-media-features-headers/#sec-ch-prefers-color-scheme

task-4282066
task-4286033
task-4812669

closes odoo/enterprise#93177

Related: odoo/odoo#224187
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>