[FIX] hr_expense_extract: prevent error when saving expense without linked user
Currently, an error occurs when creating an expense for an employee that is not
linked to a user.

Steps to reproduce:

- Install the hr_expense module.
- Create a new employee 'test' (ensure this employee is not linked to a user).
- Go to Expenses > upload [1] and open it.
- Set the Employee to 'test' and click Save (it may take up to 5 seconds).

(See [2] for Steps to reproduce)

Error:
AttributeError: 'hr.expense' object has no attribute 'uid'

Root Cause:
At [3], the code incorrectly uses self.uid instead of self.env.uid. Since
hr.expense records do not have a uid field, this leads to an error.

Fix:
This commit ensures that expenses can be saved correctly, even when the selected
employee is not linked to a user.

[1]:
https://drive.google.com/file/d/1Ew2B_zMFCQbw-6nwfKctJPy1W4y2ZaIb/view?usp=sharing

[2]:
https://drive.google.com/file/d/1qP-9L_emaox3l9JyoNVYjqj1Ht8bJSmZ/view?usp=sharing

[3]:
https://github.com/odoo/enterprise/blob/804f263b80cb78884e3f246bcc7fa7ea4e0cf73a/hr_expense_extract/models/hr_expense.py#L90

sentry-6834628909

closes odoo/enterprise#93334

X-original-commit: 50fe9c6a03e82fcaf5863f2aec281071261753a8
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>