[FIX] l10n_mx_edi: fix sat cron always checking same invoices
Currently one of the domains returned by the method
`_get_update_sat_status_domains` is `[('state', '=',
'invoice_received'), ('move_id.state', '=', 'posted')]`. This domain is
used to always return l10n_mx_edi_documents that have been imported from
somewhere and whose invoice has been posted. This is because Odoo needs
to always checked the value of the originator of an EDI document, in
case it has been cancelled from the SAT Portal for instance.

Both `state = 'invoice_received'` and `'move_id.state = 'posted'` are
mostly fixed value. The state needs to stay `invoice_received` as Odoo
needs to always check the originator document's value. And once an
invoice is posted, it's stays as so except in the case of cancellation.

This leads to an issue when the database contains more than 100
documents that are both `invoice_received` and `move_id.state =
'posted'`. In this case, the cron `_fetch_and_update_sat_status` will
always process the same 100 documents. Once the limit of 100 is reached,
the cron retriggers itself before terminating. Then on the next
execution, the search call with the domain coming from
`_get_update_sat_status_domain` will return the same 100 documents
again.

This commit fixes this issue by ordering the documents in the cron
method by `write_date asc`. Even if the SAT value of the documents does
not change, the `write_date` should be updated as their is still a write
that is triggered via `_update_document_sat_state`. This prevents the
cron from always processing the same documents over and over again.

opw-4901448

closes odoo/enterprise#97928

X-original-commit: a6c3465fb86f7af162c90e1adab7bd1da1a1c20d
Signed-off-by: Laurent Smet (las) <las@odoo.com>