[FIX] account_accountant: outstanding payments matching
Commit 4c23de148eb3689842a48df81a5ced772c214861 introduced another query to look for outstanding payments to match in the bank reco widget, aiming to
reduce the number of wrong matches found by the algorithm. Doing so, limiting the match between account.payment initiated in odoo and their matching
bank transaction imported (through stripe for example), on an exact match of the memo seemed like a good idea.

But for obscure reasons, the memo we're sending is not guaranteed to be found back, depending on the payment provider and the import flow.  Also, for
backward compatibility, it now appears important to allow the match to be on a part of the memo, like we used to do.

So we're back on a solution that splits the memo using ' - '.

closes odoo/enterprise#93237

closes odoo/enterprise#93304

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>