[FIX] l10n_nl_reports: correctly round tax closing
In this commit[^1] they reformatted the Dutch tax report, renaming the
expression labels of the lines. The functionality that does the rounding
of the amount to pay/receive depends on that label to compute the
correct rounding amount.

Currently the Dutch tax closing entry will always end up with a balance
to pay/receive of 0.0 because of that (the whole balance is "rounded").

This commit fixes that by pointing the rounding logic to the new
expression label used on the total line.

[^1]: https://github.com/odoo/odoo/commit/6078dafa68db120a3bf20e0c3e33b3886798985d

closes odoo/enterprise#93222

Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>