[FIX] pos_iot: incorrect printer used for receipts
Steps to reproduce:
1. Configure a POS with a receipt printer and a preparation printer.
2. Ensure that the preparation printer has a higher ID than the receipt
   printer (you can force this by deleting it and letting the IoT
   redetect it.)
3. Try to print a receipt.

EXPECTED: The receipt is printed on the configured receipt printer.
ACTUAL: The receipt is printed on the prepation printer.

This was caused by the hardware proxy code assuming that there will only
be 1 `iot.device` model loaded, and that it will correspond to the
receipt printer. The fix is to ensure the hardware proxy only chooses
the printer that is configured as the receipt printer.

The `pos_iot` unit test had to be adapted since it was changing the POS
config mid-test and manually calling `connectToProxy()`. The test now
uses the correct config from the start.

closes odoo/enterprise#94425

X-original-commit: 4b760dc7c061644e8fbf0fc7f85da519439567fa
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>