[IMP] account_reports: field dedicated to auto generation
Before it was clearly confusing and error prone how the auto generation works.
It was based on the field country_id on the return type. If it was set then that meant auto generated which is quite wrong as some return can have a country_id but don't want to be auto generated.

Since then we added a new field auto_generated on the return type which is by default to true. most of the return are auto generated. If we don't want a return auto generated just ticking this field to false should be enough now.

closes odoo/enterprise#93269

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>