[FIX] {pos_}appointment{_google_reserve}: error when overbooking from gantt
When using the inverse of "total reserved capacity" you could set any
number but it would be capped to the maximum capacity of all selected
resources.

Instead now if capacities are managed, warn users that they cannot book
more than the capacity of the selected resources.

This change also outlines some problematic tests that were setting
total_capacity_reserved as the inverse without setting a resource.
The tests only worked because the inverse remained set via the cache
but would have failed if the cache was invalidated at any point.

They can be safely updated to simply specify resource_ids as:
- For appointment_google_reserve, the checks were simply to ensure
that the capacity was properly forwarded. So relying on cache was fine.
- For pos, the fix commit at [1] seems to be aiming at simply making sure
the value entered in gantt is propagated "as-is" so gantt rules should be
the priority and we don't need to allow "no resources" there either.

[1]: 1e743e29bc87b67413ac36834d663253835ad589

task-4909129

Part-of: odoo/enterprise#92263
Related: odoo/odoo#222741
Related: odoo/upgrade#8262
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>