[IMP] l10n_xml_export: Modulo/Trimestre handling
Description of the issue this commit addresses:

The XML export should contain three "Modulo" sections when the report is
submitted quarterly but currently only ever sends one big chunk of data.

Also the name of the exported file is wrong.
Also the 25.82€ threshold of the vp7 line is changed for 100.00.

Desired behavior after this commit merged:

The exported file adapts to whether it is submitted monthly or quarterly and
the name of the file issue is addressed.

Community PR: https://github.com/odoo/odoo/pull/216887
task-4826511

closes odoo/enterprise#93376

X-original-commit: 3a533252cd07ddff5bfcd4b742e5e6efdf8b8b0a
Related: odoo/odoo#224587
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>