[FIX] iot: race condition when opening webrtc
If two calls are made to open a WebRTC connection within a short space
of time, they will both try to open a new connection causing a failure.
This commit fixes the issue by checking to see if a connection is
currently opening before continuing. It also adds a unit test for this
scenario.

closes odoo/enterprise#93312

Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>