[IMP] spreadsheet: untranslate list headers
Current issue:

Alice's language is in English and Bob is in French.
Alice inserts a list in a spreadsheet, and create a pivot with the dataset
being the list.
Alice chooses the pivot dimensions based on the columns headers, which are
the translated(!) field display names.

Now Bob opens the spreadsheet.
=> the pivot doesn't work because the dimension are specified using
the field names in English, but they cannot be found because they are in
French in the spreadsheet.

The issue is that `ODOO.LIST.HEADER` returns the translated field string.
One solution would be drop `ODOO.LIST.HEADER` completely and just hard-code
the string values when the list is inserted.
But users use it to add fields after, and then auto-fill it.
Also, being easily able to know which cells are list headers is helpful
to display "widget" like the sorting widget (task 4543812)

Other thing to consider: currently, when users are creating their own
spreadsheet/dashboard, there's a mix of languages: what they manually type
in cells, in their language (untranslated obviously) and the result of
ODOO.LIST.HEADER, which is translated.

Task: 4805206
Part-of: odoo/enterprise#93471
Related: odoo/odoo#224750
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>