[FIX] mrp_workorder: assign correct employee on work order completion
Issue Before This Commit:
=========================
When a work order is assigned to an employee (not to the current user),
and the current user marks it as done, the time log is wrongly created
under the current user’s employee instead of the assigned one.
This behavior caused confusion for the user as the
wrong person was shown as working on the order.

Steps to Reproduce:
===================
- Create an MO with work orders and confirm it.
- Assign another employee to a work order.
- Mark the work order as done with the current user from the work order line.
- Open the workorder and notice that the time log is created
  for the current user’s employee.

Cause of the Issue:
===================
The bug was introduced in PR (https://github.com/odoo/enterprise/pull/84790),
where the logic for assigning the main employee was overridden,
ignoring the case of an already assigned employee.

With This Commit:
=================
The time log is now created for the assigned employee,
This resolves the confusion by ensuring
the right person is tracked on the work order.

closes odoo/enterprise#100604

Taskid: 4983514
X-original-commit: 75ac3c7b26214659c90a1a495e8decfd78c47511
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Divyanshu Hasmukhbhai Joshi (jodh) <jodh@odoo.com>