[IMP] accountant_knowledge: add upload link plugin
For the audit reports, we aim to provide clear guidance on which
documents to upload in the editor to ensure a well-structured report.
To support this, we will add alert boxes in the template indicating the
required document for each section, along with a link that allows users
to upload files directly. These improvements will streamline the onboarding
process and make report creation more intuitive.

Task-4989809

closes odoo/enterprise#92223

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>