[REM] appointment: remove web_editor dependencies
This commit is part of an effort to remove dependencies on the legacy
web_editor module.

The appointment module already includes its AppointmentPlugin in the new
editor's (html_editor module) list of plugins.

This commit simply removes the code that patches the old Wysiwyg, which
is being phased out, alongside the QUnit tests for it (which already
have the equivalent ones using HOOT).

task-5046471

closes odoo/enterprise#93328

Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>