[FIX] l10n_be_reports: vatnum without country code
It could happen that the user set his vat number without the country code before
the number. In this case, we removed the two first digits of the vat number.
Also changing other occurrence using the company_vat to get the country, since
we are in the belgian ec sale list, the country_code should be 'BE' everytime

task-5039969

closes odoo/enterprise#96466

X-original-commit: b85ad6b8785c56829a050db48c9931d740bcf529
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>