[FIX] sale_subscription_stock: picking copy should update delivered qty
Before this commit, when a picking linked to a subscription is
duplicated, the delivered quantity is not updated.

The moves to use in order to compute the delivered quantity is selected
based on the date_deadline value. This allows to only account for the
moves of the current period. When a picking is duplicated, the
date_deadline and schedule date of the picking are not copied.
The moves are copied but their respective date_deadline are not.

This commit ensure that the picking and stock.move have the correct date
in order to update the delivered quantity once the picking is validated.

Moreover, chatgpt generated test_subscription_product_delivery_creation is fixed to be
understandable.

task-4910572

Part-of: odoo/enterprise#89550
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>