[FIX] pos_{restaurant_}appointment: adapt to changes in calendar views
pos_appointment inherits many views from base calendar and appointment

As we redesigned the calendar and gantt quick-edit form to use icons
we're now missing a more standard quick-edit for the context of pos
kanban Original inherited view is moved fully-compiled into POS to
maintain it. As that is the only place where it is currently needed.

Some other tweaks are done to ensure the correct views are used in POS
actions.

task-4909129

Part-of: odoo/enterprise#92263
Related: odoo/odoo#222741
Related: odoo/upgrade#8262
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>