[IMP] l10n_tr_hr_payroll: allow manual adjustments in NTG calculation
This commit allows the addition of manual deductions and additions to the net salary
when we the payslip uses a Net-To-Gross calculation.

We add a new input type and salary rule category for each of the following:
1. Manual Additions
2. Manual Deductions

We also add 3 new salary rules
1. Expected Net Salary -> this is what the NTG calculation runs on now
2. Manual Additions -> amount on the category of the same name
3. Manual Deductions -> amount on the category of the same name

and finally
(Actual) Net Salary = Expected Net - Additions Categ. + Deductions Categ.

We are using the category instead of the rule so in the future, other rules
added to the same categories work out of the box.

task-5045827

closes odoo/enterprise#93469

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>