[FIX] l10n_cl_edi: add negative price on import when `IndExe` is '6'
Issue:
When, in a DTE file, a line has the code 6 in `IndExe`. The amount of the line is supposed to be negative. (See the [doc](https://www.sii.cl/factura_electronica/formato_dte.pdf) page 31/32)

Steps to reproduce:
- Install 'l10n_cl_edi' and switrch to a Chilean company
- Have a DTE file with a line having '6' in `IndExe`:
- In Accounting > Vendor > Bills, click the button "Upload" and select the DTE file
- The imported move has positive values on the every lines

Cause:
The values of `IndExe` are not considered.

Solution:
Add a condition changing the sign of `price_unit` when 6 is in `IndExe`

opw-4954723

closes odoo/enterprise#95607

X-original-commit: 9b86dc585796b305ca8a62a64a87822e2d0a6e38
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>