[IMP] crm_helpdesk: automatically assign leads
The wizard that converts tickets to leads already has a heuristic to
assign the new leads to a user if the ticket's responsible user is
also a salesperson in the target team, but if they are not, then the
lead remains without responsible and could be missed if no automatic
assignment rules exist.

This commit extends that heuristic: the wizard now automatically assigns
the newly created lead to the sales team's leader if it failed to find
the ticket's responsible in the team's members.

task-4684660

closes odoo/enterprise#93415

Related: odoo/odoo#211855
Related: odoo/upgrade#7762
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>