[FIX] l10n_{cn_reports,fr_intrastat}: remove countries from manifest
Specifiying countries in the manifest will only auto-install the module
if at least one company is in one of the countries.

This is not wanted as, if someone already has a company with it's reports
module, and we add a new localization. We want that new localization to
auto-install it's reports module.

closes odoo/enterprise#93547

X-original-commit: bbffeb1280335cd65037f3e46c7e6e4485d044e4
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Léo Leclerc (leol) <leol@odoo.com>