[FIX] account_reports: Better company dependent default values for return types
Before the company dependent default values for the return types were set only during the ```_post_load_data``` of the chart template.
The issue with this way is that when we install a l10n for instance 10n_be_reports which add new return types, l10n_be would create a new company then trigger the _post_load_data function on return types that existed at that time. The only problem is that return types from l10n_be_reports are still not installed.

The change here is that we set the default value on the creation of the company and return type.

closes odoo/enterprise#93417

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>