[FIX] documents_hr_payroll: fix send by email button user access
Hr Payroll users that are meant to be able to use the Resend Payslip
by email button do not have enough access right to get the documents
token to put into the email "Your Payslip" button.

Add a sudo on the payslip to get the document access url after the check
of user role has been done. If have the right to use the button, sudo the rest.

Task-5049444

closes odoo/enterprise#98510

X-original-commit: 2a0b3a75866528cc499b752f74aedd3ad3629973
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: David Beguin (dbe) <dbe@odoo.com>