[FIX] l10n_ch_hr_payroll: prevented creation of new workplace via quick-create
The _rec_name was previously set to partner_id, causing type errors when typing a new workplace name in a Many2one field.
Couldn't change it to anything else, as one should have to create a new field in that stable version and this is not the best practice.
As a workaround, the corresponding field in the hr_contract_template_views.xml and hr_employee_views.xml is removed from it the quick_create to prevent the creation of new names on the fly.

task-5043056

closes odoo/enterprise#98173

X-original-commit: 062e0a61c380af2303fa70200f0f3a60529944fb
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>