[FIX] hr_work_entry_enterprise: fix work entries replace in gantt view
The code now replaces the selected cells with a new entry with the
selected type, where the duration of the new type is the sum of the
duration(s) of the previously existing entries in the corrresponding
cell. In case the cell is empty, the new entry is created according to
the expected hours of work for that day for that employee (depending
on the work entry source).

task-5043872

closes odoo/enterprise#93459

Related: odoo/odoo#224736
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>