[REF] ai_{, livechat}: replace usage of partner_id and minor fix.
In the AI app, there are multiple functions that receive the partner_id of
the ai.agent as a parameter. This complicates the code and results in
unnecessary checks on the partner_id to make sure it is connected to an AI Agent.

This commit replaces the usages of partner_id of ai.agent (unless necessary)
with the id of the agent. To do so:

1. ai_agent_id is stored on discuss.channel model.
2. All rpc calls to AI Controllers are changed to pass
ai_agent_id instead of partner_id.
3. All functions inside AI app are changed to use ai_agent_id
instead of partner_id. However, there are some functions that
actually require partner_id. These functions are left as is.

- Access to ai.agent is now managed by `_is_user_access_allowed`
method that takes into account whether the current logged in
user is internal or not and whether the ai agent can be
accessed by public/portal users.

- When ai_livechat snippet was removed, the `used_on_website_snippet`
value for the ai agent used on that snippet wasn't updated. This
commit fixes that bug.

task-4991208

Part-of: odoo/enterprise#92390
Related: odoo/odoo#223008
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>