[REV] website_sale_subscription: revert changes from one-time sale update
This commit reverts the modifications introduced in the one-time sale PR.
Those changes caused incorrect pricing behavior for products with variants, as
the necessary logic for handling multiple plans and variants was removed.

By reverting, we restore the correct price computation for products that have
both multiple plans and variants.

opw-5224319

closes odoo/enterprise#100654

X-original-commit: 82967ea05dbdee81fea3191eaf03bb9034b3d3b2
Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>