[FIX] esg: only display download button on esg database supported
Before this commit, when the user creates a new ESG database, a button
download appears on the kanban card of that database (or in the form
view). The problem is the download button will do nothing since the
structure is not uniformed for all esg database, we cannot be able to
automatically import the data of new ESG database, only the supported
ones can be downloaded (ADEME or IPCC databases).

This commit makes sure the Download button is only displayed for ADEME
and IPCC databases.

task-4812645

closes odoo/enterprise#93029

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>