[FIX] mrp_workorder: remove duplicate moves
Steps to reproduce:
1. Create a BoM with two components to be consumed in an operation
2. Create a Manufacturing Order with this BoM and confirm it
3. On the BoM, remove one of the produts and save
4. Go to the Shop Floor
5. Select the workcenter used in step 1
6. Got duplicate key in t-foreach

Before this commit:
Stock moves without quality checks, that are linked to a workorder but
not linked to a BoM line, are included twice in the view, resulting into
a traceback.

After this commit:
Include only once the stock moves, removing duplicates from the view.

Since the issue was resolved by [4a84c00], this commit only add a test
to ensure that the bug is not reintroduced.

[4a84c00]: https://github.com/odoo/enterprise/commit/4a84c00627e6473d6e8af5e193402ba593d5685c

opw-5029970

closes odoo/enterprise#96345

X-original-commit: e30087c900a301b1a482178aaaf68847c0e3b3d7
Signed-off-by: Walravens Mathieu (wama) <wama@odoo.com>