[FIX] hr_work_entry_enterprise: fix work entries split in gantt view
This commit fixes how the split button behaves in the Gantt view of the
work entry. Instead of calling the ORM function upon clicking the split
button, the split button now opens a `FormViewDialog`. The save button
in the dialog is modified so that, instead of saving the record data,
it calls the `action_split` function that handles the splitting logic.
The context is used to open the dialog with a custom form view that has
read-only fields, and the form is populated with the existing work
entry data, except for the duration, which is set as half of the
existing work entry duration.

task-5043857

closes odoo/enterprise#93449

Related: odoo/odoo#224724
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>