[IMP] ai_website_livechat: Improve UX for livechat with AI
1. add agent automatically to livechat snippet

It is easy to forget to configure an AI Agent on the AI Livechat
Snippet. So, this commit solves this problem by setting an AI Agent
automatically on the livechat snippet.

2. Restrict references shown to public users

When the LLM answers the user's query, it includes a section with all
the references used in generating the answer. In case of public users,
we don't want to show the pdf attachments in the reference section
given that public users don't have access to them and we can only show
URLs.

closes odoo/enterprise#93686

Related: odoo/odoo#225089
Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>