[IMP] knowledge: disable deletion of templates in annexe picker
The annexe picker can display a hierarchy of articles. When a user
deletes a template, both the template and its children are removed
from the database. However, while the template is removed from the
gallery, its children remain visible and clickable in the interface.
Clicking on these child templates causes the system to raise an
exception, indicating that the record does not exist.

Steps to reproduce the error:
1. Open an article created from a template.
2. Use the /index command.
3. In the index block, click on "Load a Template."
4. Delete a parent template.
5. Attempt to open a child template.
-> An error occurs stating that the record does not exist.

To prevent this issue and maintain code simplicity, this commit
disables the deletion of templates in the annexe picker.

Task-4989809

closes odoo/enterprise#93496

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>