[IMP] knowledge: defer deletion in index block
When a user deletes an article from the article index, the article is
currently permanently deleted from the database. This can be frustrating
if the user clicks the icon by mistake. To improve the experience, we
will change the behavior so that the article is moved to the trash
instead of being immediately deleted.

Additionally, when the article index becomes desynchronized, attempting
to delete an article may raise an error. To address this, we will wrap
the ORM call sending the article to the trash with a try/catch and always
resynchronize the block. This should ensure that the invalid entries are
removed from the article index.

Task-4989809

closes odoo/enterprise#93526

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>